Move comparing function to test-setup.js
